+++ src/data/niedziela-1-jutrznia.mdx
@@ -25,7 +25,8 @@
 
 <Psalm
   name="Psalm 63, 2-9"
-  antiphon={`**1 ant.** Szukam Ciebie, Boże, * pragnę ujrzeć Twą potęgę, Panie, / alleluja.`}
+  order={1}
+  antiphon="Szukam Ciebie, Boże, * pragnę ujrzeć Twą potęgę, Panie, / alleluja."
   content={[
     `Boże, mój Boże, szukam Ciebie *
 i pragnie Ciebie moja dusza.`,
